requirements.txt
streamlit
google-generativeai
FlagEmbedding
numpy
audio-recorder-streamlit
SpeechRecognition
